Update AppContent.js
The useEffect() hook allows you to run a side effect (like checkrole()) after the component has rendered. This means that the component will fully render first and then checkrole() will be called. The state updates inside checkrole() will not trigger another re-render, so the component will not get stuck in an infinite loop.